# api contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Scoop Job Seeker API
  description: |
    REST API contract for the Scoop Vue frontend.
    Base URL used by the app: `http://localhost:8000/api`
    See also `docs/API_CONTRACT.md` for human-readable handoff notes.
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /job-seeker/profile:
    get:
      tags: [Profile]
      summary: Get profile
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Profile'
    put:
      tags: [Profile]
      summary: Update profile (partial OK)
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Profile'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Profile'
...
  /jobs/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    get:
      tags: [Jobs]
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Job'
  /jobs/{id}/apply:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    post:
      tags: [Jobs]
      responses:
        '200':
          description: Applied
        '422':
          $ref: '#/components/responses/Error'
...
  /companies/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    get:
      tags: [Companies]
      security: []
      responses:
        '200':
          description: OK
  /companies/{id}/jobs:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    get:
      tags: [Companies]
      security: []
      responses:
        '200':
          description: OK
  /companies/{id}/reviews:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    get:
      tags: [Companies]
      security: []
      responses:
        '200':
          description: OK
    post:
      tags: [Companies]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                rating: { type: number }
                title: { type: string }
                body: { type: string }
      responses:
        '201':
          description: Created
...
  /tenders:
    get:
      tags: [Tenders]
      security: []
      parameters:
        - name: q
          in: query
          schema: { type: string }
      responses:
        '200':
          description: OK
  /tenders/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    get:
      tags: [Tenders]
      security: []
      responses:
        '200':
          description: OK
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        oneOf:
          - type: integer
          - type: string
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
...
  schemas:
    ErrorMessage:
      type: object
      properties:
        message: { type: string }
...
    Profile:
      type: object
      properties:
        first_name: { type: string }
        last_name: { type: string }
        email: { type: string }
        phone: { type: string }
        location: { type: string }
        gender: { type: string }
        date_of_birth: { type: string, format: date }
        employment_status: { type: string }
        highest_education: { type: string }
        driving_license: { type: boolean }
        profile_photo: { type: string, nullable: true }
        bio: { type: string }
        job_preferences:
          type: array
          items: { type: string }
        job_discovery_categories:
          type: array
          items: { type: string }
        expected_salary_min: { type: number }
        expected_salary_max: { type: number }
        joined_label: { type: string }
        profile_strength_percent: { type: integer }
        profile_sections_completed: { type: integer }
        profile_sections_total: { type: integer }
...
    Job:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        employer_name: { type: string }
        employer_id: { type: integer }
        verified: { type: boolean }
        location: { type: string }
        salary_min: { type: number }
        salary_max: { type: number }
        salary_currency: { type: string }
        category:
          type: object
          properties:
            id: { type: integer }
            name: { type: string }
        job_type:
          type: object
          properties:
            id: { type: integer }
            name: { type: string }
        positions_available: { type: integer }
        created_at: { type: string, format: date-time }
        expiry_date: { type: string, format: date-time }
        description: { type: string }
        is_saved: { type: boolean }
        application_status: { type: string, nullable: true }
        employer:
          $ref: '#/components/schemas/Employer'
    Employer:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        size: { type: string }
        industry: { type: string }
        rating: { type: number }
        reviews_count: { type: integer }
        jobs_count: { type: integer }
        location: { type: string }
        website: { type: string }
        email: { type: string }
        phone: { type: string }
        working_hours: { type: string }
        about_us: { type: string }
        logo_url: { type: string, nullable: true }
        cover_url: { type: string, nullable: true }